% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create a new project}
\usage{
create_project(
  path,
  type = c("package", "pipeline", "shinyapp"),
  cph = "Crown Copyright"
)
}
\arguments{
\item{path}{Path to directory for new project}

\item{type}{Either \code{"package"} (default) for a standard package, or one of
\code{"pipeline"} or \code{"shinyapp"}, which have a \code{run.R} file.}

\item{cph}{Copyright holder, defaults to "Crown Copyright" for work}
}
\description{
Create a new project
}
